--- src/controller_examples/zookeeper_controller/exec/zookeeper_api/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: MIT
-pub mod helper_funcs;
-pub mod api;
